Added a Page Reload When Switching Chains
Added a condition in the StateManager initialize function that will
reload the page if the user switches chains and the vaults were
previously loaded. This fixes an issue caused by the resetting of the
balances store where reactive components in the vaults page would query
properties of an empty balances store cuasing undefined errs that would
break the page.