Fixed undefined reference exceptions in reactive values on the Vaults page caused by resetting of data stores refreshing values after switching chains.
For object throwing these errs added undefined checks, and for numbers and BigNumbers
initialized values to 0 so that subsequent references to the value did
not result in additional errs that would break the page. These undefined
and default values are subsequently overwritten by an additional refresh
triggered by the fecthing of token balances which occurs right
afterwards.